THE BASICS OF MY MVC FRAMEWORK IS IN PLACE
For this project, I am reusing my pseudo-state machine idea from
tic-tac-toe. But it will now be simpler since it won't manage
form resets, etc.

And that will work because I am making page controllers for this
project. Each controller will only be concerned with items on that
"page". When a page will transition to another page, that page will
signal that intent to the viewPseudoStateMachine, and that will manage
the display of "pages".

I added:
controllers/homePage.js - the controller for the home page. Currently
it handles the button click for "sign-up" and delegates that click
to viewPseudoStateMachine. It demonstrates how the MVC framework
will work.

viewPseudoStateMachine.js - it simply shows and hides "pages" based on
the current state. Currently, this is just a shell, but it demonstrates
how things will work.

app.js - This just initializes the viewPseudoStateMachine and each
page controller.

I updated README.md. I added the flag GIFs. I wrote a description, and
I decided for this project to keep a running diary of my thoughts under
the planning section.

I chnaged the title tag from Tic Tac Toe to Welcome To Russian
Flashcards.