I IMPROVED THE MVC FRAMEWORK BY USING ES6 CLASSES
The MVC framework now uses ES6 classes. The ViewPseudoStateMachine
and HomePageController are now ES6 classes, improving modularity.

Now app.js will import CommonJS modules. This reduces the chance for
circular dependencies that I encounted in the tic-tac-toe project.
App.js now serves as the "composition root" for dependency injection.
This will promote the  "open-closed principle" and unit testability.

Home page functionality is now complete, as the button clicks delegates
to the ViewPseudoStateMachine for managing view transitions. Of course
styling is not complete, but JS functionality is, and the
ViewPseudoStateMachine doesn't actually manage views at the moment;
it simply provides hooks for doing so.