SIGNUP IS NOW FUNCTIONAL
The signup functionality is basically working with respect to
JavaScript. More may need to be done.

I created signupView.js to act as a controller for the signup page.

I updated README.md to discuss the use of classes and dependency
injection. I explained more about my MVC approach.

app.js now imports the GA provided config and store modules,
and injects that to signupView.js The config controls dev and
production URLs, and store is used to share data between controllers.

I fixed a typo in homeView.js.

I removed the alert that I had in ViewPseudoStateMachine.js