THERE IS NOW A CONTROLLER FOR STATUS MESSAGES
I created statusMessageView.js in the controllers folder. I wanted
to do this for tic-tac-toe but did not. Now there is only one piece
of JavaScript that knows how to display messages, which lets me change
how messages are displayed.

I created an injectables object in app.js, so now there is only a
single parameter for each controller. Each controller decides what it
needs. This keeps the app loosely coupled. So I had to refactor all of
the existing controllers.

I tweaked index.html so that there is an empty message, and updated the
comments.