THE VIEW PSEUDO STATE MACHINE IS WORKING
This now hides the views. It defaults to the home view. When you click
the "sign up" button, it now takes you to the sign up view. When you
successfully create an account, you are now taken to the sign in page.

Of course more views need to be added as the app development
progresses, but how it works is complete. It is also simplfied.