NAVIGATION FROM OPTIONS TO CREATE FLASHCARD WORKS
The navigation to and from create flashcard view is working. The
create flashcard view simply has a submit and return button, and the
return button can return to the options view.

I also created cyrillicKeyboardView.js. It currently does nothing, but
it is an injectable for views that might need a Cyrillic keyboard. So
I created it now so that it could be injected into the new file
createFlashcardView.js.

I obviously had to update:
- index.html
- viewPseudoStateMachine.js
- app.js
- flashcardOptionsView.js