THE CYRILLIC KEYBOARD WORKS WITH CREATE FLASHCARD
The basic Cyrillic soft keyboard is now working. The contoller for the
Cyrillic keyboard implements the Observer pattern. The controller
is an injectable to the create flashcard controller. It handles the
clicks from the soft keyboard and passes that character to each
interested controller.

I probably will add a few more keys, such as a delete key, but the
basic idea works.

I added the English and the Russian input text fields to the create
flashcard controller so that you can actually type Russian using the
soft keyboard into the input field.