A MAJOR REFACTORING OF VIEWPSEUDOSTATEMACHINE
The ViewPseudoStateMachine now has a method for controllers to register
handlers. These handlers are callbacks for functions to reset the
given view. This simplifies the WHEN in view resetting.

The method also allows a form to state what it's jQuery selector is,
so that the state machine knows the views to reset. Now the machine
subcribes to the open-closed principle.

All of the views now use the new machine logic and the old logic has
been removed.

All of the controllers have been modified to use this new machine.